+++ hutool-core/src/main/java/cn/hutool/core/util/CharUtil.java
@@ -291,6 +291,18 @@ public class CharUtil {
 				((c >= 0xE000) && (c <= 0xFFFD)) || //
 				((c >= 0x10000) && (c <= 0x10FFFF)));
 	}
+	
+	/**
+	 * 是否为Windows或者Linux（Unix）文件分隔符<br>
+	 * Windows平台下分隔符为\，Linux（Unix）为/
+	 * 
+	 * @param c 字符
+	 * @return 是否为Windows或者Linux（Unix）文件分隔符
+	 * @since 4.1.11
+	 */
+	public static boolean isFileSeparator(char c) {
+		return SLASH == c || BACKSLASH == c;
+	}
 
 	/**
 	 * 比较两个字符是否相同