+++ hutool-core/src/main/java/cn/hutool/core/util/StrUtil.java
@@ -729,6 +729,26 @@ public class StrUtil {
 	public static boolean containsAny(CharSequence str, CharSequence... testStrs) {
 		return null != getContainsStr(str, testStrs);
 	}
+	
+	/**
+	 * 查找指定字符串是否包含指定字符列表中的任意一个字符
+	 * 
+	 * @param str 指定字符串
+	 * @param testStrs 需要检查的字符数组
+	 * @return 是否包含任意一个字符
+	 * @since 4.1.11
+	 */
+	public static boolean containsAny(CharSequence str, char... testChars) {
+		if(false == isEmpty(str)) {
+			int len = str.length();
+			for(int i = 0; i < len; i++) {
+				if(ArrayUtil.contains(testChars, str.charAt(i))) {
+					return true;
+				}
+			}
+		}
+		return false;
+	}
 
 	/**
 	 * 给定字符串是否包含空白符（空白符包括空格、制表符、全角空格和不间断空格）<br>